+++ src/kotlin/webapps/service-customers/applications/src/main/kotlin/com/pal2hmnk/example/customers/infrastructures/persistence/exposed/Shops.kt
@@ -3,6 +3,6 @@ package com.pal2hmnk.example.customers.infrastructures.persistence.exposed
 import org.jetbrains.exposed.sql.Table
 
 object Shops : Table("shops") {
-    val id = Shops.long("id").autoIncrement()
+    val id = Shops.integer("id").autoIncrement()
     val name = Shops.varchar("name", 255)
 }
