+++ src/kotlin/webapps/service-customers/applications/src/main/kotlin/com/pal2hmnk/example/customers/infrastructures/persistence/exposed/Users.kt
@@ -3,6 +3,6 @@ package com.pal2hmnk.example.customers.infrastructures.persistence.exposed
 import org.jetbrains.exposed.sql.Table
 
 object Users : Table("users") {
-    val id = long("id").autoIncrement()
+    val id = integer("id").autoIncrement()
     val name = varchar("name", 255)
 }
